restructur util moved to shared
